factor containsDisallowedRune into own package (stringutils)
This will allow it to be used when receiving files, too.
