+++ fk/secretsend.go
@@ -36,6 +36,7 @@ import (
 	"github.com/fluidkeys/fluidkeys/colour"
 	"github.com/fluidkeys/fluidkeys/out"
 	"github.com/fluidkeys/fluidkeys/pgpkey"
+	"github.com/fluidkeys/fluidkeys/stringutils"
 )
 
 func secretSend(recipientEmail string, filename string) exitCode {
@@ -172,7 +173,7 @@ func scanUntilEOF() (message string, err error) {
 }
 
 func encryptSecret(secret string, filename string, pgpKey *pgpkey.PgpKey) (string, error) {
-	if containsDisallowedRune(secret) || !utf8.ValidString(secret) {
+	if stringutils.ContainsDisallowedRune(secret) || !utf8.ValidString(secret) {
 		return "", errors.New("Secret contains disallowed characters")
 	}
 
@@ -230,26 +231,3 @@ func (r *ioutilReadFilePassthrough) ReadFile(filename string) ([]byte, error) {
 }
 
 const femaleSpyEmoji = "\xf0\x9f\x95\xb5\xef\xb8\x8f\xe2\x80\x8d\xe2\x99\x80\xef\xb8\x8f"
-
-// containsDisallowedRune checks whether the given input string contains a disallowed rune
-func containsDisallowedRune(input string) bool {
-	for _, r := range input {
-		switch {
-		case 0 <= r && r <= 9:
-			return true
-		case 11 <= r && r <= 12:
-			return true
-		case r == '\n': // 10
-			continue
-		case r == '\r': // 13
-			continue
-		case 14 <= r && r <= 31:
-			return true
-		case r == 127: // DEL
-			return true
-		default:
-			continue
-		}
-	}
-	return false
-}
